--- google/cloud/grpc_utils/version_info.h.in
@@ -1,22 +1,0 @@
-// Copyright 2019 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_GRPC_UTILS_VERSION_INFO_H
-#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_GRPC_UTILS_VERSION_INFO_H
-
-#define GOOGLE_CLOUD_CPP_GRPC_UTILS_VERSION_MAJOR @GOOGLE_CLOUD_CPP_GRPC_UTILS_VERSION_MAJOR@
-#define GOOGLE_CLOUD_CPP_GRPC_UTILS_VERSION_MINOR @GOOGLE_CLOUD_CPP_GRPC_UTILS_VERSION_MINOR@
-#define GOOGLE_CLOUD_CPP_GRPC_UTILS_VERSION_PATCH @GOOGLE_CLOUD_CPP_GRPC_UTILS_VERSION_PATCH@
-
-#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_GRPC_UTILS_VERSION_INFO_H
