+++ google/cloud/samples/common_install_test.cc
@@ -12,8 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "google/cloud/grpc_utils/grpc_error_delegate.h"
 #include "google/cloud/grpc_utils/version.h"
 #include "google/cloud/version.h"
+#include <google/cloud/grpc_utils/completion_queue.h>
 #include <iostream>
 
 int main() {
@@ -23,5 +25,16 @@ int main() {
   std::cout
       << "Verify symbols from google_cloud_cpp_grpc_utils are usable. version="
       << google::cloud::grpc_utils::version_string() << "\n";
+
+  google::cloud::grpc_utils::CompletionQueue cq;
+  std::thread t{[&cq] { cq.Run(); }};
+  std::cout
+      << "Verify symbols from google_cloud_cpp_grpc_utils are usable. version="
+      << google::cloud::grpc_utils::MakeStatusFromRpcError(
+             grpc::Status(grpc::StatusCode::UNKNOWN, "Just for testing"))
+      << "\n";
+  cq.Shutdown();
+  t.join();
+
   return 0;
 }
